Refactor wagtail blocks profile directory (#3405)
* initial relocation so we can spin off individual blocks

* Update __init__.py

* refactored latest profile listing

* refactored profile directory
